pyproject.toml
[tool.poetry]
name = "ccs2snp"
version = "1.0.0"
description = "ccs2snp: SNP detection from PacBio CCS reads"
authors = ["Sangjin Lee <sl17@sanger.ac.uk>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/sjin09/ccs2snp"
repository = "https://github.com/sjin09/ccs2snp"
classifiers = [
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]

[tool.poetry.urls]
Changelog = "https://github.com/sjin09/ccs2snp/releases"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
argparse = "^1.4.0"
pysam = "^0.18.0"
biopython = "^1.78"
natsort = "^8.0.0"
numpy = "^1.20.2"
psutil = "^5.8.0"
pyfastx = "^0.8.4"
pytabix = "^0.1"
plotnine = "^0.9.0"
cyvcf2 = "^0.30.18"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
coverage = {extras = ["toml"], version = "^5.3"}
safety = "^1.9.0"
mypy = "^0.790"
typeguard = "^2.9.1"
xdoctest = {extras = ["colors"], version = "^0.15.0"}
sphinx = "^3.3.1"
sphinx-autobuild = "^2020.9.1"
pre-commit = "^2.8.2"
flake8 = "^3.8.4"
black = "^20.8b1"
flake8-bandit = "^2.1.2"
flake8-bugbear = "^20.1.4"
flake8-docstrings = "^1.5.0"
flake8-rst-docstrings = "^0.0.14"
pep8-naming = "^0.11.1"
darglint = "^1.5.8"
reorder-python-imports = "^2.3.6"
pre-commit-hooks = "^3.3.0"
sphinx-rtd-theme = "^0.5.0"
sphinx-click = "^2.5.0"
Pygments = "^2.7.2"

[tool.poetry.scripts]
ccs2snp = "ccs2snp.__main__:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["ccs2snp"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
